requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings>=2.2.1
openpyxl==3.1.5
python-docx==1.1.2
